Deleted vgg16 model files en replaced that with a notebook. Also made a file to hold the functions that been used at all classification models.
